requirements.txt
requests
Pillow
spotipy
pixoo
six
python-dotenv
